refactor: 대시보드 역할별 query와 header 분리
