+++ src/features/dashboard/components/parent/index.tsx
@@ -9,8 +9,8 @@ import {
   useParentDashboardInquiryListQuery,
   useParentDashboardStudyConsultationQuery,
   useParentDashboardStudyNewsQuery,
-} from '../../hooks/use-dashboard-query';
-import DashboardHeader from '../header';
+} from '../../hooks/use-parent-dashboard-query';
+import ParentDashboardHeader from '../header/parent-header';
 import { ClassConsultationHistorySection } from '../section/parent-class-consultation-history-section';
 import { ConsultationSection } from '../section/parent-consultation-section';
 import { ParentLinkSection } from '../section/parent-link-section';
@@ -139,7 +139,7 @@ const DashboardParent = () => {
 
   return (
     <div className="flex w-full flex-col">
-      <DashboardHeader />
+      <ParentDashboardHeader />
       <main className="tablet:gap-12 desktop:gap-20 bg-gray-white tablet:py-12 desktop:pb-25 tablet:px-20 relative flex w-full flex-col gap-8 px-4.5 py-8">
         <ParentLinkSection connectedStudents={connectedStudentData ?? []} />
         <div className="tablet:gap-25 flex w-full flex-col gap-8">
