+++ src/features/dashboard/components/section/teacher-qna-section.tsx
@@ -1,50 +1,46 @@
 'use client';
 
-import { useAuth } from '@/features/auth/hooks/use-auth';
-import { useStudentDashboardQnaListQuery } from '@/features/dashboard/hooks/use-student-dashboard-query';
+import { useTeacherDashboardQnaListQuery } from '@/features/dashboard/hooks/use-teacher-dashboard-query';
 import { trackDashboardQnaMoreClick } from '@/shared/lib/analytics';
 
 import QnASectionContent from '../section-content/qna-section-content';
 import DashboardSection from './single-section';
 
-type Props = {
+type TeacherQnASectionProps = {
   className?: string;
 };
 
-const QnASection = ({ className }: Props) => {
-  const { member } = useAuth();
-  const { data: studentQnaData, isPending } = useStudentDashboardQnaListQuery({
+const TeacherQnASection = ({ className }: TeacherQnASectionProps) => {
+  const { data, isPending } = useTeacherDashboardQnaListQuery({
     page: 0,
     size: 3,
     sortKey: 'LATEST',
-    enabled: member?.role === 'ROLE_STUDENT',
   });
-
-  const questions = studentQnaData?.content ?? [];
 
   return (
     <DashboardSection
-      title="나의 질문"
+      title="답변이 필요한 질문"
+      description="아직 답변하지 않은 질문만 추렸어요."
       className={className}
       isMore={true}
       isMoreHref="/dashboard/qna"
       isMorePrefetch={false}
-      onMoreClick={() => trackDashboardQnaMoreClick(member?.role)}
+      onMoreClick={() => trackDashboardQnaMoreClick('ROLE_TEACHER')}
     >
       {isPending ? (
         <div className="flex w-full flex-col gap-1">
-          {Array.from({ length: 3 }).map((_, i) => (
+          {Array.from({ length: 3 }).map((_, index) => (
             <div
-              key={i}
+              key={index}
               className="bg-gray-3 h-12 w-full animate-pulse rounded-lg"
             />
           ))}
         </div>
       ) : (
-        <QnASectionContent questions={questions} />
+        <QnASectionContent questions={data?.content ?? []} />
       )}
     </DashboardSection>
   );
 };
 
-export default QnASection;
+export default TeacherQnASection;
