+++ src/features/dashboard/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useDashboardQuery } from './use-dashboard-query';
 export * from './use-student-note-timer';
